Persist host list state between runs
If the proxy is restarted after a crash, it should pick up with the same
host list as it had before. To achieve this we write the list of hosts
to a file whenever it changes. On startup, if the file is present we
read our initial list of hosts back from it.