Parse ENV vars for run config
Previously we were translating them in the `Dockerfile`, but that was
requiring running the command in a shell, which interferes with signals.
Also it means it only works when running in Docker.

Instead, we can check the environment variables when setting up the
default values.

We allow them to be set with or without a prefix, to avoid name clashes.
For example, `--http-port` can be set in the environment as either
`HTTP_PORT` or `MPROXY_HTTP_PORT`.