Add `pause` and `resume` commands
This adds the ability to pause an active target. When paused, new
requests will block, before being passed to the upstream server. Once
the target is resumed, the blocked requests will be passed to the
upstream server as usual.

This effectively allows us to queue up requests before processing them,
which is useful for operations like database failover that cause
upstream servers to be unable to respond successfully for a short window
of time.