Allow limiting pause duration
When pausing, accept a `--max-pause` option that specifies how long each
request can remain paused before it is cancelled.

The duration is measured form the start of each request, not the pause
time. In this way, if the paused runs for longer than the max time,
we'll shed requests oldest-first. The most recent requests are the ones
most likely to still work, so those are the ones to keep.