Claim requests within lock to avoid race condition
There was a race condition previously, where a target might be switched
from `Active` to `Draining` while we were serving a request from it. If
this happens after we fetch the active target, but before we start the
inflight request, then the request would find the target in the wrong
state and reject it.

To avoid this, we split the work of `ServeHTTP` into two parts:

- `StartRequest` sets up the in-flight request. We can do this while we
  have the lock. Once a request has been started, the target will be
  able to serve it even if it changes to draining.
- `SendRequest` actually serves the request that we previously started.
  We can do this part outside of the lock.