Request buffering
Adds the ability to buffer requests before forwarding them to the target
server. Buffering is not enabled by default, but can be controlled
through two new deploy options:

 - `buffer-requests` - set this flag to enable buffering (default is
   false)
 - `buffer-memory` - max size (in bytes) to buffer into memory, when
   buffering is enabled. Defaults to 1MB. Once this limit is exceeded
   for a request, the remainder of its body be buffered to a tempfile on
   disk.

The existing `max-request-body` still limits the total size of a request
body. When buffering is enabled, this effectively limits the size of the
tempfiles that will be created when exceeding buffer memory.
`max-request-body` now defaults to 1GB. Requests with bodies larger than
`max-request-body` will result in a `413 (Request Entity Too Large)`
response.