Allow logging custom request & response headers
You can now specify headers to log from the request and/or the response,
using a combination of `--log-request-header` and
`--log-response-header`.

To add multiple headers, you can specify the flags multiple times, or
use a comma-delimited list. For example:

    --log-response-header=Cache-Control,X-Revision

Header names are converted to snake case in the output, and prefixed
with `req` or `resp`. So the headers from the above example will be
logged as `resp_cache_control` and `resp_x_revision`.