Render proxy errors as HTML
Adds simple HTML error pages for proxy errors like 502, 503, etc.

This is partly because they look a bit nicer, but also so that people
could add branded error pages if they want. This could be useful when
using `kamal-proxy stop` to take an app offline for maintenance, and so
on.

We don't (currently) provide a flag for setting where these errors come
from, but they could be replaced by mounting a folder into the location
we serve them from (/usr/local/share/kamal-proxy/pages).