Overwrite existing socket file if necessary
Previously a stale socket file would cause the server to start up
without a command handler. The HTTP(S) listeners would start, but the
command handler would silently fail to start. This would leave us with a
running server that we had no way of interacting with.

Instead, we'll remove any stale socket if there's one in the way.
This lets us more easily recover from a previous crash that left the
file behind. We don't remove the file until after successfully starting
the HTTP(S) listeners, so that helps to guard against accidentally
running the same server twice and losing the first socket.

Additionally, now if the command handler fails to start we'll exit the
server with an error. Better to bail out than proceed in a half-working
state.