Reject requests over TLS if it has been disabled
If we were running TLS on a site and then disable it, any active
connections that were made over TLS will still be valid until closed.
This could be confusing if you disable TLS then refresh your browser and
find that it's still working.

Instead, we can reject the request with an error so that the change
applies to active connections as well.