Buffered responses shouldn't implicitly add status
If the wrapped handler completes without writing anything, then
previously we were implicitly sending a 200 status when writing the
heeader.

In most cases this wouldn't matter, since a 200 is the default. However,
an outer middleware might interpret that empty response differently, and
then need to set a different status. Which it can't do if we've already
written one.

Specifically in our case, there's an outer error page middleware which
may need to generate an error response that the innermost handler
triggered before returning an empty response. If the buffered response
adds a 200 of its own, it prevents the error page middleware from being
able to write the correct status later.