Forward headers by default if not terminating TLS
In most cases, `--tls` will correspond with whether we're receiving
traffic directly, or are sitting behind another load balancer or proxy.

This means we can infer a default for `--forward-headers` that is more
likely to match what we want:

  - If we're handling TLS, don't forward the incoming headers as they
    are may be from untrusted clients.
  - If we're *not* handling TLS, something downstream of us probably is.
    In which case, we probably want to forward its headers.

We can still specify the option directly with `--forwarwd-headers`. This
only changes the default behaviour when that flag is not set, so that
omitting it is more likely to do what's expected.