Allow routing to wildcard subdomains
This allows routing to wildcard subdomains by specifying a target like:

    --target=*.example.com

This matches requests to hosts with any subdomain in place of the `*`.
It does *not* match multiple levels of subdomain, nor are multiple `*`
allowed in a single host; it behaves much like wildcard matching in TLS
certificates.

Automatic TLS is not currently supported with wildcard hosts (because
the challenge types we use -- `http-01` and `tls-alpn-01` -- can't be
used to generate wildcard certificates). However you can bring your own
wildcard certificates using `--tls-certificate-path` /
`--tls-private-key-path`.