Don't flush responses while buffering
In order to make SSE responses work properly, the underlying response
writers all need to be flushable (otherwise responses may not be sent
immediately when they should be). When our response buffering is
enabled, we bypass our buffer for SSE responses for the same reason.

However, if we are buffering a non-SSE response and our reponse writer's
`Flush` is called, we need to avoid actually flushing anything to the
underlying writer. Otherwise we would send the response too early,
rather than buffering it as intended. Worse, if we can `Flush` on the
underlying writer without having first sent it the buffered content,
then what gets flushed will be wrong: we'll send an empty response
rather than the real response.

To fix this, we should ignore calls to `Flush` when we're buffering.