Introduce path based routing
Adds a new `--path-prefix` option for deployments. When specified, only
requests beginning with that prefix will be routed to the service. This
allows multiple service to share the same hosts.

`--path-prefix` needn't be a single path segment, but the prefix must
split on a path boundary. For example, a prefix of `/api/internal` is
valid, and will match a path of `/api/internal/users/123`, or
`/api/internal`; but it would not match `/api/internal2`.

By default, the prefix will be removed before forwarding the request on
to the target. So, given a `--path-prefix=/api`, a request to
`/api/users` will reach the target as `/users`.

In the case where the prefix should be preserved, we can set
`--strip-path-prefix=false`. In this case, the full path is preserved.