Ensure locks released before draining old target
Previously, we were holding onto the service & target locks during a
deployment while the draining stage was in progress. This can cause
requests to be blocked until the drain completes.

Instead, ensure that the locks are released as soon as we're finished
with them, and before starting to drain the old target.