Ensure locks are released before draining
Previously, we were holding onto the service & target locks during a
deployment while the draining stage was in progress. This can cause
requests to be blocked until the drain completes.

Instead, we should ensure that the locks are release as soon as we're
finished with them, and we can let the drain continue after that.