`tls-disable-redirect` -> `tls-redirect`
Turning off the redirect previously involved setting its `disabled`
value to `true`, which works fine but in retrospect feels a little like
a double negative. Instead we can have a `tls-redirect` option that
defaults to `true`, but can be set to `false` if we want to turn the
behaviour off.