Allow splitting reads/writes when load balancing
In addition to specifying one or more `--target` servers to deploy, you
can now also specify one or more `--read-target`. When any read targets
are present, traffic will be routed such that read requests go to read
targets, and write requests go to write targets.

A read request is defined as a `GET` or `HEAD`, and excludes WebSocket
traffic. Everything else is considered to be a write request.

Read requests that immediately follow a write on the same session will
also be treated as writes. In configurations where readers are replicas
of writers, this helps avoid problems where replication lag leads to a
client failing to read an item it just wrote. This "writer affinity" is
controlled by a cookie, and by default the effect lasts for 3 seconds.
It can be adjusted via `--writer-affinity-timeout`, or disabled entirely
by setting the timeout to zero.

Aside from this writer affinity case, writers will not receive any read
requests unless they are also explicitly listed as a `--read-target`