Skip writer affinity via header
In cases where a write should be excluded from triggering writer
affinity, a client can set the header `X-Writer-Affinity=false`. This
can be useful for certain requests, like beacons, that result in
unnecessary pinning to the writer because the result of those requests
isn't visible on the page (meaning that the risk of reading stale data
after writing doesn't apply).