Support writer affinity times < 1s
Since the cookie expiration is in whole seconds, setting an affinity
timeout of less that 1s will cause the cookie to expire too soon. The
cookie value gives us the precision we need to have sub-1s affinity
timeouts, but we need to ensure the cookie lasts long enough to read it.

To solve this we can round up the cookie expiry to the next second.