Allow pinning writer with `X-Writer`
If a response returns an `X-Writer` header, we will use the value of
that header as the target to send future write requests too. If that
target is not available, future writes will result in a 503.

Currently we'll only process this header if we're the outermost proxy
for the request. This is to prevent problems when proxies are nested
(e.g. a load balancer proxy that routes to individual Kamal-deployed
hosts, each with their own proxy for gapless deployments).