Re-proxying
This adds a "re-proxying" feature. Re-proxying is a bit like an internal
redirect. An upstream may respond to a request with a new URL to reproxy
to. That URL can specify the hostname of a different upstream within the
same load balancer, and/or a different path. This URL should be returned
in the `X-Kamal-Reproxy-Target` header.

When the proxy receives a non-success response that contains a URL in
that header, it will re-submit the request to that URL, using the same
method and body content as the original request. The hostname in the URL
will be matched against the list of targets currently deployed into the
service.

Re-proxying is only available when the `--dynamic-load-balancing` option
is enabled. When enabled, requests will have a `X-Kamal-Reproxy: true`
header added, so that upstream servers can tell whether reproxying is
available.