Update PauseController to support jsonv2
Our previous mechanism for avoiding infinite recursion in our custom
`UnmarshalJSON` relied on the fact that the `json` implementation would
not call `UnmarshalJSON` on the aliased pointer type. However this is
not the case in `jsonv2`; it will call it, and thus results in infinite
recursion.

To handle this situation correctly with `jsonv2` we should introduce a
valid type alias instead.