Add `--cookie-path-prefix`
Adds a mechanism to automatically scope cookie paths to match
`--path-prefix`.

When `--cookie-path-prefix` is `true` and a request matches a
`--path-prefix`, any cookies set by the response will automatically have
their paths scoped to the matched prefix.

For example, given a service that is deployed with:

    ... --path-prefix=/app,api

And a response that sets a cookie like:

    Set-Cookie: session=1; Secure; Path=/; Domain=example.com

Then for a request to `/app/something`, the client will actually receive
a cookie header of:

    Set-Cookie: session=1; Secure; Path=/app; Domain=example.com

This feature is intended to make it easier to allow multiple, distinct
applications to share the same domain by using different paths, without
them accidentally clobbering each others' cookies.