Serve a default cert if SNI server name missing
Generally we expect all TLS traffic to use SNI, and that allows us to
match up the connection to the deployment that holds its cert.

However, some scenarios might require making a TLS request without SNI.
(One specific case we've seen for this is a healthcheck from an F5 load
balancer.) So to have a best-effort attempt at servicing these requests,
we'll default to using the first available TLS-enabled hostname that we
have if there is no SNI to go off.