Use cookie for writer pinning
Rather that using a header to pin the current writer, use a cookie. This
allows the writer to be different for different clients, without the
proxy needing to be involved in how that's organised.

Also, if we try to pin, or reproxy, to a writer that is not available
(doesn't exist, or is unhealthy) then try another writer instead. This
should make it easier for a session to heal when the set of healthy
writers changes.