Ignore missing targets when pin/reproxy
When pinning to a target with a cookie, or reproxying to a target, then
if that target doesn't exist we should pick a different one at random.
This gives a better chance of healing if a target goes unhealthy (since
one of the other healthy targets ought to be able to either handle it,
or reproxy it to wherever else can).