Add reproxy delay
When reproxying over many attempts, add a delay between them. This is to
help cover the case where an upstream server is temporarily unavailable,
or settings are being changed that will affect the final routing.

We'll do the first 3 attempts as fast as we can, but then introduce a 1
second delay between each attempt until either we get a successful
response, or the maximum 30 retries have run out.

To make this more useful, we'll also only reproxy to the specified host
if it is healthy. If it's not, we'll try picking another from the list.