Add `--default-writer` for dynamic routing
Normally when there are multiple writers, we'll load balance writes
between them. But when using dynamic routing to control the writes with
a cookie, we might want to specify a default writer to use for the case
where the cookie is not set.

The `--default-writer` option specifies which writer to use for all
writes that aren't directed by a cookie. If `--default-writer` is empty
(or, if dynamic routing is not enabled) then we'll load balance writes
between writers as before.