Add `--force` for deployments
When using Kamal Proxy as a load balancer with multiple targets, it can
sometimes be useful to update a configuration even when some of those
targets are not yet responding (for example, during maintenance or
outage).

To allow this, add a `--force` switch on the `deploy` comamnd. When
`--force` is set, we won't want for targets to become healthy before
applying the new configuration.