Use `bypass` to avoid double header write
We can get the same behaviour as [f11cf9af] by leaning on the existing
`bypass` flag, which tracks that fact that we've switched to
unbuffered.

Also don't think the test is carrying its weight here. It's really just
a reminder of the fact that a warning used to be emitted here, but the
external behaviour hasn't changed.